requirements.txt
livekit-agents[openai,deepgram,silero]
